feat(textarea): add start and end slots (#28441)
Issue number: Part of #26297

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

With the modern form control syntax, it is not possible to add icon
buttons or other decorators to the sides of `ion-textarea`, as you can
with `ion-item`.

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

`start` and `end` slots added. While making this change, I also tweaked
the CSS selectors responsible for translating the label above the input
with `"stacked"` and `"floating"` placements, merging this logic into a
single class managed by the TSX. I needed to add a new class for whether
slot content is present, so the selectors were getting unwieldy
otherwise.

Docs PR TBA; I plan on knocking out all three components at once when
the features are all complete, to make dev builds easier to manage.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this introduces a breaking change, please describe the impact
and migration path for existing applications below. -->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->

---------

Co-authored-by: ionitron <hi@ionicframework.com>
Co-authored-by: Maria Hutt <thetaPC@users.noreply.github.com>